Finish examples refactor
