+++ src/Passes/Source/Profiles/BaseProfile.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#include "Passes/Profile/Profile.hpp"
+#include "ProfilePass/Profile.hpp"
 #include "Profiles/BaseProfile.hpp"
 #include "Rules/Factory.hpp"
 #include "Rules/RuleSet.hpp"
